complete

+++ Car.cpp
@@ -103,7 +103,7 @@ void Car::run1() {
 	}// here we are stopped somewhere in the drop off zone
 	cout << "Stopped. Front: " << front << " Unloading..." << endl;
 	speed = SPEED0;
-	hold(abs(normal(unloadTime, 3.5)));
+	hold(abs((int)normal(unloadTime, 3.5)));
 	cout << "...Done. Front: " << front << " Departing." << endl;
 	
 	while( front < roadZone->getEnd() ) { // while index of the front of the car is < index of the end
@@ -259,7 +259,7 @@ void Car::run2() {
 	}// here we are stopped somewhere in the drop off zone
 	cout << "Stopped. Front: " << front << " Unloading..." << endl;
 	speed = SPEED0;
-	hold(abs(normal(unloadTime, 3.5)));
+	hold(abs((int)normal(unloadTime, 3.5)));
 	cout << "...Done. Front: " << front << " Departing." << endl;
 	
 	while( front < roadZone->getEnd() ) { // while index of the front of the car is < index of the end
